+++ utils/__init__.py
@@ -1,4 +1,5 @@
-# __all__ = ["helpers", "constants"]
+# __all__ = ["helpers", "constants", "log_helper"]
 
 from .helpers import *
 from .constants import *
+from .log_helper import *